+++ core/model/src/main/java/com/sangtq/model/forecast/Hour.kt
@@ -23,7 +23,7 @@ data class Hour(
     val temp_c: Double = 0.0,
     val temp_f: Double = 0.0,
     val time: String = "",
-    val time_epoch: Int = 0,
+    val time_epoch: Long = 0,
     val uv: Int = 0,
     val vis_km: Int = 0,
     val vis_miles: Int = 0,
